Refactor, get some subpages sorta kinda working
Okay, still ugly as sin, but version _AND_ speedtest are basically
working.  Speedtest is still fragile as hell due to `omitempty` issue
for integers => nil in json decoding (WTF).  Need a more intelligent
solution.

Templating is improved and we have exposed the strings.Join() method
to templates.

Refactored the server code to be easier to manage.